requirements.txt
httpx
pyyaml
loguru
watchdog
nest_asyncio
whatsapp_chatbot_python